refactor: checking sed version (#70)
This PR address this issue which asks for a graceful exit in cases where the version of sed is less than 4.3. I have made this version check in my commit and used exit 1 if the check fails. Do let me know if any changes are required. Specifically, I am unsure about the position of the check inside the code and if the exit needs to be more comprehensive (like killing child processes if they exist, etc.).